Refactored code to be in a trait.
