+++ src/Patches.php
@@ -18,16 +18,18 @@ use Composer\Package\AliasPackage;
 use Composer\Package\PackageInterface;
 use Composer\Plugin\PluginInterface;
 use Composer\Installer\PackageEvents;
+use Composer\Repository\InstalledRepositoryInterface;
+use Composer\Repository\RepositoryManager;
 use Composer\Script\Event;
 use Composer\Script\ScriptEvents;
 use Composer\Installer\PackageEvent;
 use Composer\Util\ProcessExecutor;
 use Composer\Util\RemoteFilesystem;
-use RecursiveArrayIterator;
-use RecursiveIteratorIterator;
 use Symfony\Component\Process\Process;
 
 class Patches implements PluginInterface, EventSubscriberInterface {
+
+  use PatchesIgnoreTrait;
 
   /**
    * @var Composer $composer
@@ -53,6 +55,22 @@ class Patches implements PluginInterface, EventSubscriberInterface {
    * @var array $installedPatches
    */
   protected $installedPatches;
+  /**
+   * @var RepositoryManager $repositoryManager
+   */
+  protected $repositoryManager;
+  /**
+   * @var InstalledRepositoryInterface $localRepository
+   */
+  protected $localRepository;
+  /**
+   * @var PackageInterface $packages
+   */
+  protected $packages;
+  /**
+   * @var array $patches_temp_list
+   */
+  protected $patches_temp_list;
 
   /**
    * Apply plugin modifications to composer
@@ -67,6 +85,10 @@ class Patches implements PluginInterface, EventSubscriberInterface {
     $this->executor = new ProcessExecutor($this->io);
     $this->patches = array();
     $this->installedPatches = array();
+    $this->repositoryManager = $composer->getRepositoryManager();
+    $this->localRepository = $this->repositoryManager->getLocalRepository();
+    $this->packages = $this->localRepository->getPackages();
+    $this->patches_temp_list = array();
   }
 
   /**
@@ -91,7 +113,10 @@ class Patches implements PluginInterface, EventSubscriberInterface {
 
   /**
    * Before running composer install,
+   *
    * @param Event $event
+   *
+   * @throws \Exception
    */
   public function checkPatches(Event $event) {
     if (!$this->isPatchingEnabled()) {
@@ -99,23 +124,21 @@ class Patches implements PluginInterface, EventSubscriberInterface {
     }
 
     try {
-      $repositoryManager = $this->composer->getRepositoryManager();
-      $localRepository = $repositoryManager->getLocalRepository();
       $installationManager = $this->composer->getInstallationManager();
-      $packages = $localRepository->getPackages();
 
       // Gather up the extra for the main composer.json
       $extra = $this->composer->getPackage()->getExtra();
       $patches_ignore = isset($extra['patches-ignore']) ? $extra['patches-ignore'] : array();
-      $full_patches_ignore = $patches_ignore;
-      $tmp_patches = $this->grabPatches();
-      $full_patches = $tmp_patches;
+      $this->patches_temp_list = $this->grabPatches();
+
+      $this->full_patch_list = $this->patches_temp_list;
+      $this->full_patch_ignore_list = $patches_ignore;
 
       // First Pass: Walk through packages against main composer.json
-      foreach ($packages as $package) {
+      foreach ($this->packages as $package) {
         $extra = $package->getExtra();
         if (isset($extra['patches'])) {
-          $full_patches = $this->arrayMergeRecursiveDistinct($full_patches, $extra['patches']);
+          $this->full_patch_list = $this->arrayMergeRecursiveDistinct($this->full_patch_list, $extra['patches']);
           // Apply the main composer.json patches-ignore list
           if (isset($patches_ignore[$package->getName()])) {
             foreach ($patches_ignore[$package->getName()] as $package_name => $patches_to_ignore) {
@@ -124,61 +147,37 @@ class Patches implements PluginInterface, EventSubscriberInterface {
               }
             }
           }
-          unset($patches_to_ignore);
           $this->installedPatches[$package->getName()] = $extra['patches'];
         }
         $patches = isset($extra['patches']) ? $extra['patches'] : array();
-        $tmp_patches = $this->arrayMergeRecursiveDistinct($tmp_patches, $patches);
+        $this->patches_temp_list = $this->arrayMergeRecursiveDistinct($this->patches_temp_list, $patches);
       }
-      if (is_array($tmp_patches)) {
+      if (is_array($this->patches_temp_list)) {
         $this->io->write('<info>First Pass: main composer.json patches collected.</info>');
-        $this->writePatchLog('patches-full', $full_patches, 'Full list of patches in all packages');
+        $this->writePatchLog('patches-full', $this->full_patch_list, 'Full list of patches in all packages');
       }
 
-      // Second Pass: Remove patches from patches-ignore in dependencies
-      if ($this->isPackagePatchingEnabled()) {
-        $installedPatches = $this->installedPatches;
-        foreach ($packages as $package) {
-          $extra = $package->getExtra();
-          // Apply the package patches-ignore list
-          if ($this->checkPatchesIgnoreLegal($package) && isset($extra['patches-ignore'])) {
-            $package_patches_ignore = $extra['patches-ignore'] ?? [];
-            $full_patches_ignore = $this->arrayMergeRecursiveDistinct($full_patches_ignore, $package_patches_ignore);
-            foreach ($this->flattenPatchesIgnore($package_patches_ignore) as $package_name => $patches_to_ignore) {
-              if (isset($installedPatches[$package->getName()][$package_name])) {
-                $installedPatches[$package->getName()][$package_name] = array_diff($installedPatches[$package->getName()][$package_name], $patches_to_ignore);
-              }
-            }
-            unset($patches_to_ignore);
-          }
-          $patches = isset($extra['patches']) ? $extra['patches'] : array();
-          $tmp_patches = $this->arrayMergeRecursiveDistinct($tmp_patches, $patches);
-        }
-        $this->installedPatches = $installedPatches;
-        $this->io->write('<info>Second Pass: dependency composer.json patches ignored.</info>');
-        $this->writePatchLog('patches-ignore', $full_patches_ignore, 'Full list of patches to ignore from all packages');
-        $this->writePatchLog('patches', $installedPatches, 'Final list of patches to be applied');
-      }
+      $this->doPatchesIgnoreCollation();
 
-      if ($tmp_patches == FALSE) {
+      if ($this->patches_temp_list == FALSE) {
         $this->io->write('<info>No patches supplied.</info>');
         return;
       }
 
       // Remove packages for which the patch set has changed.
       $promises = array();
-      foreach ($packages as $package) {
+      foreach ($this->packages as $package) {
         if (!($package instanceof AliasPackage)) {
           $package_name = $package->getName();
           $extra = $package->getExtra();
-          $has_patches = isset($tmp_patches[$package_name]);
+          $has_patches = isset($this->patches_temp_list[$package_name]);
           $has_applied_patches = isset($extra['patches_applied']) && count($extra['patches_applied']) > 0;
           if (($has_patches && !$has_applied_patches)
             || (!$has_patches && $has_applied_patches)
-            || ($has_patches && $has_applied_patches && $tmp_patches[$package_name] !== $extra['patches_applied'])) {
+            || ($has_patches && $has_applied_patches && $this->patches_temp_list[$package_name] !== $extra['patches_applied'])) {
             $uninstallOperation = new UninstallOperation($package, 'Removing package so it can be re-installed and re-patched.');
             $this->io->write('<info>Removing package ' . $package_name . ' so that it can be re-installed and re-patched.</info>');
-            $promises[] = $installationManager->uninstall($localRepository, $uninstallOperation);
+            $promises[] = $installationManager->uninstall($this->localRepository, $uninstallOperation);
           }
         }
       }
@@ -486,112 +485,6 @@ class Patches implements PluginInterface, EventSubscriberInterface {
   }
 
   /**
-   * Checks if the root package enables sub-package patching.
-   *
-   * @return bool
-   *   Whether sub-package patching is enabled. Defaults to TRUE.
-   */
-  protected function isPackagePatchingEnabled() {
-    $extra = $this->composer->getPackage()->getExtra();
-
-    if (empty($extra['patches']) && empty($extra['patches-ignore']) && !isset($extra['patches-file'])) {
-      return $extra['enable-patches-ignore-subpackages'] ?? FALSE;
-    }
-    else {
-      return TRUE;
-    }
-  }
-
-  /**
-   * Checks to see if we have any Patches to Ignore from the whitelist.
-   *
-   * @return bool
-   *   Whether the whitelist exists or not.
-   */
-  protected function checkPatchesIgnoreWhitelist() {
-    $extra = $this->composer->getPackage()->getExtra();
-
-    if (empty($extra['patches-ignore-whitelist'])) {
-      return FALSE;
-    } else {
-      return TRUE;
-    }
-  }
-
-  /**
-   * Checks to see if the patches being ignored are in fact legal to use.
-   *
-   * @param \Composer\Package\PackageInterface $package
-   *
-   * @return bool
-   */
-  protected function checkPatchesIgnoreLegal(PackageInterface $package) {
-    if (!$this->isPackagePatchingEnabled()) {
-      return FALSE;
-    }
-
-    $extra = $package->getExtra();
-    $package_patches_ignore = $extra['patches-ignore'] ?? [];
-    $package_name = $package->getName();
-    if ($this->checkPatchesIgnoreWhitelist()) {
-      $patches_package_whitelist = $this->getPatchesIgnoreWhitelist();
-      if (in_array($package_name, $patches_package_whitelist)) {
-        return TRUE;
-      } else {
-        return FALSE;
-      }
-    }
-    if (!$this->checkPatchesIgnoreWhitelist() && isset($package_patches_ignore)) {
-      return TRUE;
-    }
-    return FALSE;
-  }
-
-  /**
-   * flattenPatchesIgnore returns the lowest leaf of a multidimensional array.
-   *
-   * If you have
-   *
-   * @param $package_patches_ignore
-   *
-   * @return array
-   */
-  protected function flattenPatchesIgnore($package_patches_ignore) {
-    if ($this->isMultidimensionalArray($package_patches_ignore)) {
-      foreach($package_patches_ignore as $package_name => $patches) {
-        if ($this->isMultidimensionalArray($patches)) {
-          $this->flattenPatchesIgnore($patches);
-        } else {
-          return [$package_name => $patches];
-        }
-      }
-    }
-    if (isset($patches)) {
-      return $patches;
-    }
-    return $package_patches_ignore;
-  }
-
-
-  protected function getPatchesIgnoreWhitelist() {
-    $extra = $this->composer->getPackage()->getExtra();
-    return $extra['patches-ignore-whitelist'];
-  }
-
-  /**
-   * Checks to see if the requested array is multidimensional or not.
-   *
-   * @param array $array
-   *   The array to check.
-   * @return bool
-   *   TRUE or FALSE return.
-   */
-  protected function isMultidimensionalArray(array $array):bool
-  {
-    return is_array($array[array_key_first($array)]);
-  }
-
-  /**
    * Writes a patch report to the target directory.
    *
    * @param array $patches
@@ -605,89 +498,6 @@ class Patches implements PluginInterface, EventSubscriberInterface {
       $output .= 'Source: ' . $url . "\n\n";
     }
     file_put_contents($directory . "/PATCHES.txt", $output);
-  }
-
-  /**
-   * Writes a patch log for debugging purposes.
-   *
-   * @param array $patches
-   * @param string $directory
-   */
-  protected function writePatchLog($filename, array $patches, $message = "Patches applied to this folder") {
-    $package = $this->composer->getPackage();
-
-    if ($this->checkPatchLog()) {
-      $directory = $this->getPatchLogParameter('location');
-      $file_format = strtolower($this->getPatchLogParameter('format'));
-      if (!file_exists($directory)) {
-        mkdir($directory, 0755);
-      }
-      switch ($file_format) {
-        case 'yml':
-        case 'yaml':
-          file_put_contents($directory . "/" . $filename . ".yml", yaml_emit($patches, JSON_PRETTY_PRINT));
-          break;
-        case 'json':
-          file_put_contents($directory . "/" . $filename . ".json", json_encode($patches, JSON_PRETTY_PRINT));
-          break;
-        case 'txt':
-        case 'text':
-          $output = $message . ":\n\n";
-          foreach ($patches as $patch => $patch_info) {
-            if (isset($patch_info) && !empty($patch_info)) {
-              if ($this->isMultidimensionalArray($patch_info)) {
-                $output .= '=== ' . $patch;
-                foreach ($this->flattenPatchesIgnore($patch_info) as $flatten_patch => $info) {
-                  $output .= ' ' . $flatten_patch . "\n";
-                  foreach ($info as $description => $url) {
-                    $output .= '   ' . $url . "\n";
-                    $output .= '   ' . $description . "\n\n";
-                  }
-                }
-              } else {
-                $output .= '=== ' . $patch . "\n";
-                foreach ($patch_info as $description => $url) {
-                  $output .= '   ' . $url . "\n";
-                  $output .= '   ' . $description . "\n\n";
-                }
-              }
-            }
-          }
-          file_put_contents($directory . "/" . $filename . ".txt", $output);
-
-          // TODO: Add a collapsed set of files here.
-
-          break;
-        case 'php':
-        default:
-          // Raw output
-          file_put_contents($directory . "/" . $filename . ".php", var_export($patches, true));
-      }
-    }
-  }
-
-  /**
-   * @return bool
-   */
-  protected function checkPatchLog() {
-    $extra = $this->composer->getPackage()->getExtra();
-    if (isset($extra['patches-log'])) {
-      return TRUE;
-    }
-    return FALSE;
-  }
-
-  /**
-   * @param $parameter
-   *
-   * @return false|mixed
-   */
-  protected function getPatchLogParameter($parameter) {
-    $extra = $this->composer->getPackage()->getExtra();
-    if (isset($extra['patches-log'][$parameter])) {
-      return $extra['patches-log'][$parameter];
-    }
-    return FALSE;
   }
 
   /**
@@ -752,29 +562,6 @@ class Patches implements PluginInterface, EventSubscriberInterface {
     }
 
     return $merged;
-  }
-
-  /**
-   * A recursive search for a needle in an array haystack.
-   *
-   * @param $needle
-   * @param array $haystack
-   *
-   * @return mixed
-   */
-  public function recursiveSearch($needle, array $haystack)
-  {
-    $iterator  = new RecursiveArrayIterator($haystack);
-    $recursive = new RecursiveIteratorIterator(
-      $iterator,
-      RecursiveIteratorIterator::SELF_FIRST
-    );
-    foreach ($recursive as $key => $value) {
-      if ($key === $needle) {
-        return TRUE;
-      }
-    }
-    return FALSE;
   }
 
   /**
